+++ oura-library/src/main/kotlin/org/radarbase/oura/converter/OuraDailyOxygenSaturationConverter.kt
@@ -23,7 +23,7 @@ class OuraDailyOxygenSaturationConverter(
                 TopicData(
                     key = user.observationKey,
                     topic = topic,
-                    offset = startInstant.toEpochMilli().toDouble(),
+                    offset = startInstant.toEpochMilli(),
                     value = it.toDailySpo2(startInstant),
                 )
             }
