+++ oura-library/src/main/kotlin/org/radarbase/oura/converter/OuraSleepHrvConverter.kt
@@ -43,7 +43,7 @@ class OuraSleepHrvConverter(
                     TopicData(
                         key = user.observationKey,
                         topic = topic,
-                        offset = time,
+                        offset = time.toLong(),
                         value = toHrv(
                             time,
                             timeReceivedEpoch,
